+++ src/ts/Html5Time.ts
@@ -1,3 +1,5 @@
+import * as Conviva from '@convivainc/conviva-js-coresdk';
+
 export class Html5Time implements Conviva.TimeInterface {
   public getEpochTimeMs(): number {
     const d = new Date();