+++ src/ts/Html5Timer.ts
@@ -1,8 +1,7 @@
-import TimerCancelFunction = Conviva.TimerCancelFunction;
-import TimerAction = Conviva.TimerAction;
+import * as Conviva from '@convivainc/conviva-js-coresdk';
 
 export class Html5Timer implements Conviva.TimerInterface {
-  public createTimer(timerAction: TimerAction, intervalMs: number, actionName?: string | null): TimerCancelFunction {
+  public createTimer(timerAction: Conviva.TimerAction, intervalMs: number, actionName?: string | null): Conviva.TimerCancelFunction {
     let timerId = window.setInterval(timerAction, intervalMs);
     return function () {
       if (timerId !== -1) {